+++ NWN.Anvil/src/main/API/Object/CreatureClassInfo.cs
@@ -43,7 +43,7 @@ namespace Anvil.API
     /// <param name="spellLevel">The spell level for the spell to be added.</param>
     public void AddKnownSpell(NwSpell spell, byte spellLevel)
     {
-      classInfo.AddKnownSpell(spellLevel, (uint)spell.Id);
+      classInfo.AddKnownSpell(spellLevel, spell.Id);
     }
 
     /// <summary>
@@ -52,7 +52,7 @@ namespace Anvil.API
     /// <param name="spell">The spell to clear.</param>
     public void ClearMemorizedKnownSpells(NwSpell spell)
     {
-      classInfo.ClearMemorizedKnownSpells((uint)spell.Id);
+      classInfo.ClearMemorizedKnownSpells(spell.Id);
     }
 
     /// <summary>
@@ -128,7 +128,7 @@ namespace Anvil.API
     /// <param name="spell">The spell to remove.</param>
     public void RemoveKnownSpell(byte spellLevel, NwSpell spell)
     {
-      classInfo.RemoveKnownSpell(spellLevel, (uint)spell.Id);
+      classInfo.RemoveKnownSpell(spellLevel, spell.Id);
     }
 
     /// <summary>
